refactor(core): extract analysis workflow owner
